Fixing functionality of the codet
